refactor task token service
