+++ src/app/modules/item/http-services/save-grade.service.ts
@@ -4,7 +4,7 @@ import { Observable } from 'rxjs';
 import { appConfig } from 'src/app/shared/helpers/config';
 import * as D from 'io-ts/Decoder';
 import { decodeSnakeCase } from 'src/app/shared/operators/decode';
-import { TaskToken } from './generate-task-token.service';
+import { TaskToken } from './task-token.service';
 import { TaskScore, TaskScoreToken } from '../task-communication/types';
 import { AnswerToken } from './generate-answer-token.service';
