+++ src/app/modules/item/http-services/task-token.service.ts
@@ -5,34 +5,32 @@ import { map } from 'rxjs/operators';
 import * as D from 'io-ts/Decoder';
 import { appConfig } from 'src/app/shared/helpers/config';
 import { decodeSnakeCase } from 'src/app/shared/operators/decode';
+import { ActionResponse, successData } from 'src/app/shared/http-services/action-response';
 
-export const taskTokenResponseDecoder = D.struct({
-  data: D.struct({
-    taskToken: D.string,
-  }),
-  message: D.string,
-  success: D.boolean,
+const taskTokenDataDecoder = D.struct({
+  taskToken: D.string,
 });
 
-type TaskTokenResponse = D.TypeOf<typeof taskTokenResponseDecoder>;
-export type TaskToken = TaskTokenResponse['data']['taskToken'];
+type TaskTokenData = D.TypeOf<typeof taskTokenDataDecoder>;
+export type TaskToken = TaskTokenData['taskToken'];
 
 @Injectable({
   providedIn: 'root',
 })
-export class GenerateTaskTokenService {
+export class TaskTokenService {
 
   constructor(private http: HttpClient) {}
 
   generate(itemId: string, attemptId: string, asTeamId?: string): Observable<TaskToken> {
     const params = asTeamId ? new HttpParams({ fromObject: { as_team_id: asTeamId } }) : undefined;
-    return this.http.post<unknown>(
+    return this.http.post<ActionResponse<unknown>>(
       `${appConfig.apiUrl}/items/${itemId}/attempts/${attemptId}/generate-task-token`,
       undefined,
       { params },
     ).pipe(
-      decodeSnakeCase(taskTokenResponseDecoder),
-      map(({ data }) => data.taskToken),
+      map(successData),
+      decodeSnakeCase(taskTokenDataDecoder),
+      map(data => data.taskToken),
     );
   }
